# Conflicts:
#	src/main/java/com/backend/digitalhouse/ClinicaOdontologica/dto/entrada/turno/TurnoEntradaDto.java
#	src/main/java/com/backend/digitalhouse/ClinicaOdontologica/dto/salida/turno/TurnoSalidaDto.java
#	src/main/java/com/backend/digitalhouse/ClinicaOdontologica/repository/TurnoRepository.java
#	src/main/java/com/backend/digitalhouse/ClinicaOdontologica/service/impl/TurnoService.java

+++ src/main/java/com/backend/digitalhouse/ClinicaOdontologica/dto/salida/turno/TurnoSalidaDto.java
@@ -3,7 +3,7 @@ package com.backend.digitalhouse.ClinicaOdontologica.dto.salida.turno;
 import java.time.LocalDateTime;
 
 public class TurnoSalidaDto {
-    private Long id;
+    private int id;
     private PacienteTurnoSalidaDto pacienteTurnoSalidaDto;
     private OdontologoTurnoSalidaDto odontologoTurnoSalidaDto;
     private LocalDateTime fechaYHora;
@@ -11,18 +11,18 @@ public class TurnoSalidaDto {
     public TurnoSalidaDto(){
     }
 
-    public TurnoSalidaDto(Long id, PacienteTurnoSalidaDto pacienteTurnoSalidaDto, OdontologoTurnoSalidaDto odontologoTurnoSalidaDto, LocalDateTime fechaYHora) {
+    public TurnoSalidaDto(int id, PacienteTurnoSalidaDto pacienteTurnoSalidaDto, OdontologoTurnoSalidaDto odontologoTurnoSalidaDto, LocalDateTime fechaYHora) {
         this.id = id;
         this.pacienteTurnoSalidaDto = pacienteTurnoSalidaDto;
         this.odontologoTurnoSalidaDto = odontologoTurnoSalidaDto;
         this.fechaYHora = fechaYHora;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
